# docs/openapi.yml
openapi: '3.1.0'
info:
  title: Civic Influence Graph API
  version: 0.1.0
  description: |
    RESTful JSON API for the Civic Influence Graph platform.
    All responses include provenance metadata. No identity-based targeting filters
    are exposed. Authentication required only for write operations and bulk exports.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Default API base path

# ------------------------------------------------------------------
# Common components
# ------------------------------------------------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API key passed as Bearer token
  schemas:
    PaginationMeta:
      type: object
      properties:
        request_id: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        data_snapshot: { type: string, nullable: true, format: date-time }
        total_count: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
...
    SavedQuery:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        query_type: { type: string, enum: [graph, search, table] }
        query_config: { type: object }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
paths:
...
  /saved-queries:
    post:
      summary: Create a saved query
      tags: [Saved Queries]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, query_type, query_config]
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                query_type: { type: string, enum: [graph, search, table] }
                query_config: { type: object }
      responses:
        '201':
          description: Saved query created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/SavedQuery' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        '401': { description: Unauthorized }
        '400': { description: Validation error }
  /saved-queries/{id}:
    get:
      summary: Retrieve a saved query
      tags: [Saved Queries]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Saved query data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/SavedQuery' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        '404': { description: Not found }
  /saved-queries/{id}/execute:
    get:
      summary: Re-execute a saved query against current data
      tags: [Saved Queries]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Execution results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      saved_query: { $ref: '#/components/schemas/SavedQuery' }
                      results: { type: object }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        '404': { description: Not found }
...